CRUD functionality with working Angular UI.

+++ frontend/src/app/admincontroller/filetypeextensions/filetypeextensions.component.ts
@@ -1,16 +1,18 @@
 import { Component, OnInit } from '@angular/core';
 import { FileService } from 'src/app/servers.service';
-
+import { CanFiletypeExtensionsComponentDeactivate } from './can-deactivate-filetype-extn.service';
+import { Observable } from 'rxjs';
 
 @Component({
   selector: 'app-filetypeextensions',
   templateUrl: './filetypeextensions.component.html',
   styleUrls: ['./filetypeextensions.component.css']
 })
-export class FiletypeextensionsComponent implements OnInit {
+export class FiletypeextensionsComponent implements OnInit, CanFiletypeExtensionsComponentDeactivate {
   fileTypeExtensions = [];
   successMsgDiv = 'hide';
   deleteObj;
+  changesSaved = true;
 
   constructor(private fileService: FileService) { }
 
@@ -42,19 +44,27 @@ export class FiletypeextensionsComponent implements OnInit {
     }
   }
   addFiletype() {
+    this.changesSaved = false;
     this.fileTypeExtensions.unshift({ id: null, extensions: null, file_type: null, file: true, edit: true });
   }
-  removeExtension(extension) {
-    let cnfrm = window.confirm("Are you sure you want to delete ?")
-    let index = this.fileTypeExtensions.indexOf(extension);
-    if (cnfrm == true) {
+  removeExtension(data) {
+    if (data.file_type == null || data.extensions == null) {
+      let index = this.fileTypeExtensions.indexOf(data);
       this.deleteObj = this.fileTypeExtensions.splice(index, 1)[0];
-     
-      this.fileService.deleteFiletypeExtensions(this.deleteObj).subscribe(
-        (response) => console.log(response),
-        (error) => console.log(error)
-      )
+      this.changesSaved = true;
+    }else{
+      let cnfrm = window.confirm("Are you sure you want to delete ?")
+      let index = this.fileTypeExtensions.indexOf(data);
+      if (cnfrm == true) {
+        this.deleteObj = this.fileTypeExtensions.splice(index, 1)[0];
+
+        this.fileService.deleteFiletypeExtensions(this.deleteObj).subscribe(
+          (response) => console.log(response),
+          (error) => console.log(error)
+        )
+      }
     }
+
   }
 
   // processExtensions() {
@@ -106,10 +116,20 @@ export class FiletypeextensionsComponent implements OnInit {
 
   onSaveClick() {
     // Right now we are just pushing the top value of the stack.
+    console.log(this.fileTypeExtensions[0]);
     this.fileService.createFiletypeExtensions(this.fileTypeExtensions[0]).subscribe(
       (response) => {
         console.log(response);
       }, (error) => console.log(error)
     );
+  }
+
+  canDeactivate(): Observable<boolean> | Promise<boolean> | boolean {
+    if (!(this.changesSaved)) {
+      return confirm("Do you want to discard the changes?");
+    }
+    else {
+      return true;
+    }
   }
 }
